Remove unsupported things by Stainless.
